Moved some reusable code from cc65/coptstop.c into new files.
